Moved login_form.dart to new folder, and refactored code to reflect changes in other dart files.
